finish refactoring tests
